refactor: add LeaderSchedule trait
