+++ poh/src/poh_recorder.rs
@@ -1935,9 +1935,11 @@ mod tests {
         slot_leaders.extend(std::iter::repeat(leader_d_pubkey).take(consecutive_leader_slots));
         let mut leader_schedule_cache = LeaderScheduleCache::new_from_bank(&bank);
         let fixed_schedule = solana_ledger::leader_schedule::FixedSchedule {
-            leader_schedule: Arc::new(
-                solana_ledger::leader_schedule::LeaderSchedule::new_from_schedule(slot_leaders),
-            ),
+            leader_schedule: Arc::new(Box::new(
+                solana_ledger::leader_schedule::CurrentLeaderSchedule::new_from_schedule(
+                    slot_leaders,
+                ),
+            )),
         };
         leader_schedule_cache.set_fixed_leader_schedule(Some(fixed_schedule));
